Poison Upgrade

+++ Assets/HealthBar.cs
@@ -7,14 +7,14 @@ public class HealthBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
-        slider.maxValue = health;
-        slider.value = health;
+        slider.maxValue = (int)health;
+        slider.value = (int)health;
     }
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
-        slider.value = health;
+        slider.value = (int)health;
     }
 }
